refactor: 하위 component props로 setCalculateInfo 넘겨주는 로직 제거
