+++ calculator/src/utils.js
@@ -1,0 +1,3 @@
+export const convertToLocaleString = (number) => {
+  return number.toLocaleString("ko-KR");
+};
